completed first efforts to make it usable
there's still a bug that hasn't been fixed, when the user moves to the
next question without answering the previous the timer for the previous
question continues to run.

the ideal behavior would be to have a way of remembering the time
remaining for a particular question. that way if the user skips a
question and comes back to it later on, the timer continues from where
it left off.