fixed cors problems
it took 3 hours to fix this... the issue is that browsers do not allow
fetch calls to external servers unless the headers of the requested
resource have the `AccessControlAllowOrigin` header.

I still have to research this more, in order to understand how to
properly manage this property. for now the quick-fix is to allow
requests from anywhere.