re-formatted to use 4 spaces instead of 2 and more
I also realized that some properties where not being updated as
intended. for instance, the score value would remain the same, even
though the user might have already switch to a different question. this
basically meant that the maximum score that the user could get would be
the score of the first question multiplied by the number of questions.
which obviously does not necessarily add up to 100%.