added score adjusting feature
basically, I didn't want to have to manually calculate how much should
each question weight in order to add up to 100%. plus, there might be
cases where the users asks for `n` questions. calculating the weight for
such cases beforehand is impossible, so now the server takes care of
that for us.