added feature that shuffles flashcard array
if you do multiple sessions in a day, with little time in between them
you will probably remember the answers to some questions because you
know the order in which they appeared. this makes sure that every time
you start a new session the question are in a different order.