fixed memory leak bug
inside the `CircularTimer` component two `useEffect`
calls where made:

    1. read the `isMounted.current` property
    2. change the value of `isMounted.current`

this lead to a problem because React relies on the order
in which hooks are called. which means that sometimes the
read being done in (1) was outdated, which lead to a memory
leak because it was mutating the state after the component
had been unmounted.