fixed infinte renders bug
this one was basically due to the fact of having circular dependencies
inside the `useEffect` hook. basically, the `FlashCardAnswer` component
and the `FlashCardQuestion` components change on every render. since
they were used inside a `useEffect` hook they triggered the `useEffect`
to run on ever render. which of coursed caused the component to
re-render and thereby changed the `FlashCardQuestion` and
`FlashCardAnswer` component again, causing an infinte number of renders.

the fix was basically to remove the `useEffect` call that relied on both
components.