cleaned up interface
before I relied on the user being able to write Markdown so I let the
page title of `Answer` come from the markdown written by the user. So
the user had to write something like `## Answer` on every answer... I
later realized this was stupid, so I hardcoded the `Answer` titlepage as
an `h2` header.